Bug Fixes
- Fixed an issue where division operations were not producing correct results.

New Features
- Added support for floating-point operations.
- Implemented handling of multi-digit numbers.

+++ stack.h
@@ -1,18 +1,19 @@
 #ifndef STACK_H
 #define STACK_H
 
-#define MAX 100 // Maximum size of the stack in RPN calculator
+#define MAX 100 // Maximum size of the stack
 
+// Stack structure to store double values
 typedef struct
 {
-    int data[MAX];
+    double data[MAX]; // Store double instead of int for correct division results
     int top;
 } Stack;
 
 // Function declarations
 void initStack(Stack *s);
-void push(Stack *s, int value);
-int pop(Stack *s);
-int peek(Stack *s);
+void push(Stack *s, double value);
+double pop(Stack *s);
+double peek(Stack *s);
 
 #endif
